move util funcs to proper module
