+++ chord_metadata_service/restapi/api_views.py
@@ -1,18 +1,22 @@
 import logging
 
-from collections import Counter, defaultdict
-from typing import Tuple, Mapping
-from calendar import month_abbr
+from collections import Counter
 
 from django.conf import settings
 from django.views.decorators.cache import cache_page
-from django.db.models import Count, F, Func, CharField, IntegerField, Case, When, Value
-from django.db.models.functions import Cast
 from rest_framework.permissions import AllowAny
 from rest_framework.response import Response
 from rest_framework.decorators import api_view, permission_classes
 
-from chord_metadata_service.restapi.utils import parse_individual_age
+from chord_metadata_service.restapi.utils import (
+    parse_individual_age,
+    stats_for_field,
+    compute_binned_ages,
+    get_field_bins,
+    get_categorical_stats,
+    get_date_stats,
+    get_range_stats
+)
 from chord_metadata_service.chord.permissions import OverrideOrSuperUserOnly
 from chord_metadata_service.metadata.service_info import SERVICE_INFO
 from chord_metadata_service.chord import models as chord_models
@@ -291,237 +295,3 @@ def public_overview(_request):
         response["fields"][field]["data"] = stats
 
     return Response(response)
-
-
-def get_categorical_stats(field_props):
-    """
-    Fetches statistics for a given categorical field and apply privacy policies
-    """
-    model, field_name = get_model_and_field(field_props["id"])
-    stats = stats_for_field(model, field_name, add_missing=True)
-
-    # Enforce values order from config and apply policies
-    threshold = settings.CONFIG_PUBLIC["rules"]["count_threshold"]
-    labels: list[str] = field_props["config"]["enum"]
-    bins = []
-
-    for category in labels:
-        v = stats.get(category, 0)
-        if v and v < threshold:
-            v = 0
-        bins.append({"label": category, "value": v})
-
-    if stats["missing"] > 0:
-        bins.append({"label": "missing", "value": stats["missing"]})
-
-    return bins
-
-
-def get_date_stats(field_props):
-    """
-    Fetches statistics for a given date field, fill the gaps in the date range
-    and apply privacy policies.
-    Note that dates within a JSON are stored as strings, not instances of datetime.
-    TODO: for now, only dates in extra_properties are handled. Handle dates as
-    regular fields when needed.
-    TODO: for now only dates binned by month are handled
-    """
-    LENGTH_Y_M = 4 + 1 + 2  # dates stored as yyyy-mm-dd
-    threshold = settings.CONFIG_PUBLIC["rules"]["count_threshold"]
-
-    if field_props["config"]["bin_by"] != "month":
-        msg = f"Binning dates by `{field_props['config']['bin_by']}` method not implemented"
-        logger.error(msg)
-        raise NotImplementedError(msg)
-
-    model, field_name = get_model_and_field(field_props["id"])
-
-    if "extra_properties" not in field_name:
-        msg = "Binning date-like fields that are not in extra-properties is not implemented"
-        logger.error(msg)
-        raise NotImplementedError(msg)
-
-    query_set = model.objects.all()\
-        .values(field_name)\
-        .order_by(field_name)\
-        .annotate(total=Count(field_name))   # Note: lexical sort works on ISO dates
-
-    stats = defaultdict(int)
-    start = end = None
-    # Key the counts on yyyy-mm combination (aggregate same month counts)
-    for item in query_set:
-        key = 'missing' if item[field_name] is None else item[field_name][:LENGTH_Y_M]
-        stats[key] += item["total"]
-
-        if key == 'missing':
-            continue
-        if start:
-            end = key
-        else:
-            start = key
-
-    # All the bins between start and end date must be represented
-    bins = []
-    if start:   # at least one month
-        for year, month in monthly_generator(start, end or start):
-            key = f"{year}-{month}"
-            label = f"{month_abbr[month].capitalize()} {year}"    # convert key as yyyy-mm to `abbreviated month yyyy`
-            v = stats.get(key, 0)
-            bins.append({
-                "label": label,
-                "value": 0 if v < threshold else v
-            })
-
-    # Append missing items at the end if any
-    if 'missing' in stats:
-        bins.append({"label": "missing", "value": stats["missing"]})
-
-    return bins
-
-
-def get_range_stats(field_props):
-    threshold = settings.CONFIG_PUBLIC["rules"]["count_threshold"]
-    model, field = get_model_and_field(field_props["id"])
-
-    # Generate a list of When conditions that return a label for the given bin.
-    # This is equivalent to an SQL CASE statement.
-    whens = [When(**{f"{field}__gte": floor},  **{f"{field}__lt": ceil}, then=Value(label))
-             for floor, ceil, label in labelled_range_generator(field_props)]
-
-    query_set = model.objects\
-        .values(label=Case(*whens, default=Value("missing"), output_field=CharField()))\
-        .annotate(total=Count("label"))
-
-    stats: Mapping[str, int] = dict()
-    for item in query_set:
-        key = item["label"]
-        stats[key] = item["total"] if item["total"] > threshold else 0
-
-    # All the bins between start and end must be represented and ordered
-    bins = []
-    for floor, ceil, label in labelled_range_generator(field_props):
-        bins.append({"label": label, "value": stats.get(label, 0)})
-
-    if "missing" in stats:
-        bins.append({"label": "missing", "value": stats["missing"]})
-
-    return bins
-
-
-def labelled_range_generator(field_props) -> Tuple[int, int, str]:
-    """
-    Note: limited to operations on integer values for simplicity
-    A word of caution: when implementing handling of floating point values,
-    be aware of string format (might need to add precision to config?) computations
-    of modulo and lack of support for ranges.
-    """
-
-    c = field_props["config"]
-    minimum = int(c["minimum"])
-    maximum = int(c["maximum"])
-    taper_left = int(c["taper_left"])
-    taper_right = int(c["taper_right"])
-    bin_size = int(c["bin_size"])
-
-    # check prerequisites
-    # Note: it raises an error as it reflects an error in the config file
-    if maximum < minimum:
-        raise ValueError(f"Wrong min/max values in config: {field_props}")
-
-    if (taper_right < taper_left
-            or minimum > taper_left
-            or taper_right > maximum):
-        raise ValueError(f"Wrong taper values in config: {field_props}")
-
-    if (taper_right - taper_left) % bin_size:
-        raise ValueError(f"Range between taper values is not a multiple of bin_size: {field_props}")
-
-    # start generator
-    if minimum != taper_left:
-        yield minimum, taper_left, f"{minimum}-{taper_left}"
-
-    for v in range(taper_left, taper_right, bin_size):
-        yield v, v + bin_size, f"{v}-{v + bin_size}"
-
-    if maximum != taper_right:
-        yield taper_right, maximum, f"≥ {taper_right}"
-
-
-def monthly_generator(start: str, end: str) -> Tuple[int, int]:
-    """
-    generator of tuples (year nb, month nb) from a start date to an end date
-    as ISO formated strings `yyyy-mm`
-    """
-    [start_year, start_month] = [int(k) for k in start.split("-")]
-    [end_year, end_month] = [int(k) for k in end.split("-")]
-    last_month_nb = (end_year - start_year) * 12 + end_month
-    for month_nb in range(start_month, last_month_nb):
-        year = start_year + month_nb // 12
-        month = month_nb % 12 or 12
-        yield year, month
-
-
-def get_model_and_field(field_id: str) -> Tuple[any, str]:
-    model_name, *field_path = field_id.split("/")
-
-    if model_name == "individual":
-        model = pheno_models.Individual
-    elif model_name == "experiment":
-        model = experiments_models.Experiment
-    else:
-        msg = f"Accessing field on model {model_name} not implemented"
-        logger.error(msg)
-        raise NotImplementedError(msg)
-
-    field_name = "__".join(field_path)
-    return model, field_name
-
-
-def stats_for_field(model, field: str, add_missing=False) -> Mapping[str, int]:
-    # values() restrict the table of results to this COLUMN
-    # annotate() creates a `total` column for the aggregation
-    # Count() aggregates the results by performing a GROUP BY on the field
-    query_set = model.objects.all().values(field).annotate(total=Count(field))
-    stats: Mapping[str, int] = dict()
-    for item in query_set:
-        key = item[field]
-        if key is None:
-            continue
-
-        key = key.strip()
-        if key == "":
-            continue
-
-        stats[key] = item["total"]
-
-    if add_missing:
-        isnull_filter = {f"{field}__isnull": True}
-        stats['missing'] = model.objects.all().values(field).filter(**isnull_filter).count()
-
-    return stats
-
-
-def get_field_bins(model, field, bin_size):
-    # creates a new field "binned" by substracting the modulo by bin size to
-    # the value which requires binning (e.g. 28 => 28 - 28 % 10 = 20)
-    # cast to integer to avoid numbers such as 60.00 if that was a decimal,
-    # and aggregate over this value.
-    query_set = model.objects.all().annotate(
-        binned=Cast(
-            F(field) - Func(F(field), bin_size, function="MOD"),
-            IntegerField()
-        )
-    ).values('binned').annotate(total=Count('binned'))
-    stats = {item['binned']: item['total'] for item in query_set}
-    return stats
-
-
-def compute_binned_ages(bin_size):
-    a = pheno_models.Individual.objects.filter(age_numeric__isnull=True).values('age')
-    binned_ages = []
-    for r in a.iterator():  # reduce memory footprint (no caching)
-        if r["age"] is None:
-            continue
-        age = parse_individual_age(r["age"])
-        binned_ages.append(age - age % OVERVIEW_AGE_BIN_SIZE)
-    return binned_ages
